Move <ConceptIcon /> to common/
