+++ app/javascript/components/modals/exercise-completed-modal/ConceptProgression.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Icon } from '../../common'
-import { ConceptIcon } from './ConceptIcon'
+import { ConceptIcon } from '../../common/ConceptIcon'
 
 const ProgressBar = ({
   isMastered,
